Move rules builder interface into separate file
